types: add global and shims for app and nitro
