+++ packages/app/src/index.ts
@@ -1,3 +1,2 @@
 export * from './nuxt'
-export * from './shim'
 export * from './composables'
