+++ packages/app/types/index.d.ts
@@ -1,0 +1,3 @@
+import './shims'
+
+export * from '../dist/app'
