+++ packages/nitro/src/index.ts
@@ -2,5 +2,4 @@ export * from './build'
 export * from './context'
 export * from './server/middleware'
 export * from './server/dev'
-export * from './types'
 export { wpfs } from './utils/wpfs'
