--- packages/nitro/src/types.ts
@@ -1,11 +1,0 @@
-import type { $Fetch } from 'ohmyfetch'
-
-declare global {
-  const $fetch: $Fetch
-
-  namespace NodeJS {
-    interface Global {
-      $fetch: $Fetch
-    }
-  }
-}
