+++ packages/nitro/types/index.d.ts
@@ -1,0 +1,3 @@
+import './shims'
+
+export * from '../dist/index'
