+++ tsconfig.json
@@ -10,7 +10,9 @@
     "resolveJsonModule": true,
     "types": [
       "node",
-      "jest"
+      "jest",
+      "@nuxt/app",
+      "@nuxt/nitro",
     ]
   },
   "exclude": ["./packages/*/dist/*"]
